Drowsiness_Detection
A study by the Central Road Research Institute (CRRI) observes that about 40% of road accidents occurs due to exhausted drivers who doze off at the wheel. These statistics indicate the significance of a reliable drowsiness detection system to inform the driver before a collision occurs. Our answer to this problem is to create a detection system that recognises key characteristics of sleepiness and sends an alarm when someone becomes drowsy before it is too late. S P RAHUL KUMAR and I have implemented the drowsiness detection system using OpenCV and dlib in python.